[python, api] Fix bugs, add tests.
Run mypy and fix problems it found.

- Add and fix various tests in it2api and test_it2api.sh
- Handle None in various places
- Fix improper override of async_get in PartialProfile. Renamed that method to async_query. async_get will return a full profile. Updated it2api to use it.
- Update darknight and theme examples for new APIs.
- Add grid_size to Session
- Cancel tasks when Connection.async_main returns to avoid warnings that happen sometimes. I was able to reproduce the warning by creating a window whose session immediately exited. The error was that a task was still pending.
- Move constants in FocusUpdateWindowChanged to an Enum.
- Remove some debug print statemsn in focus.py
- Fix a typo in Profile.set_cursor_color that broke it
- Fix screen streamer when requesting screen contents.
- Fix WindowedCoordRange.right
- Fix fetching screen contents, which always returned nothing because PTYSession computed the range improperly.

+++ api/library/python/iterm2/iterm2/color.py
@@ -36,37 +36,37 @@ class Color:
         return self.__red
 
     @red.setter
-    def red(self, value: float):
+    def red(self, value: int):
         """Sets the color's red component."""
         self.__red = value
 
     @property
-    def green(self) -> float:
+    def green(self) -> int:
         """The color's green component."""
         return self.__green
 
     @green.setter
-    def green(self, value: float):
+    def green(self, value: int):
         """Sets the color's green component."""
         self.__green = value
 
     @property
-    def blue(self) -> float:
+    def blue(self) -> int:
         """The color's blue component."""
         return self.__blue
 
     @blue.setter
-    def blue(self, value: float):
+    def blue(self, value: int):
         """Sets the color's blue component."""
         self.__blue = value
 
     @property
-    def alpha(self) -> float:
+    def alpha(self) -> int:
         """The color's alpha component."""
         return self.__alpha
 
     @alpha.setter
-    def alpha(self, value: float):
+    def alpha(self, value: int):
         """Sets the color's alpha component."""
         self.__alpha = value
